+++ src/telliot_feeds/reporters/tips/funded_feeds/call_functions.py
@@ -1,6 +1,5 @@
 from typing import Any
 from typing import Callable
-from typing import List
 from typing import Optional
 
 from multicall import Call
@@ -119,7 +118,7 @@ class CallFunctions(AssembleCall):
         self,
         feed_id: bytes,
         query_id: bytes,
-        timestamps: Optional[List[int]],
+        timestamps: Optional[list[int]],
         handler_function: Optional[Callable[..., Any]] = None,
     ) -> Call:
         """getRewardClaimStatusList autopay Call object for list of timestamps
@@ -140,4 +139,43 @@ class CallFunctions(AssembleCall):
             feed_id=feed_id,
             query_id=query_id,
             param3=timestamps,
+        )
+
+    def get_current_feeds(self, query_id: bytes, handler_function: Optional[Callable[..., Any]] = None) -> Call:
+        """getCurrentFeeds autopay Call object for a query id
+
+        Args:
+        - query_id
+        - handler_function: function to interpret contract reponse (optional)
+
+        Return: Call object
+        example return from this call:
+        >>> {b'query_id': [b'feed_id',]}
+        """
+        return self.assemble_call_object(
+            func_sig="getCurrentFeeds(bytes32)(bytes32[])",
+            returns=[
+                [query_id, handler_function],
+            ],
+            query_id=query_id,
+        )
+
+    def get_data_feed(self, feed_id: bytes, handler_function: Optional[Callable[..., Any]] = None) -> Call:
+        """getDataFeed autopay Call object for month old timestamp
+
+        Args:
+        - feed_id
+        - handler_function: function to interpret contract reponse (optional)
+
+        Return: Call object
+        example return from this call:
+        >>> {b'feed_id': (reward, balance, startTime, interval, window, priceThreshold, rewardIncreasePerSecond,
+        feedsWithFundingIndex)}
+        """
+        return self.assemble_call_object(
+            func_sig="getDataFeed(bytes32)((uint256,uint256,uint256,uint256,uint256,uint256,uint256))",
+            returns=[
+                [feed_id, handler_function],
+            ],
+            feed_id=feed_id,
         )
